d-ai-async/api-results/requirements.txt
fastapi
uvicorn
python-dotenv
azure-identity
azure-cosmos
opentelemetry-api
opentelemetry-sdk
azure-monitor-opentelemetry
azure-core-tracing-opentelemetry
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
python-multipart